+++ knapsack_library/src/models/knapsack_solver.rs
@@ -4,5 +4,5 @@ use crate::models::knapsack::Knapsack;
 pub trait KnapsackSolver {
     fn get_name(&self) -> String;
 
-    fn solve(&self, knapsack: &Knapsack) -> u32;
+    fn solve(&self, knapsack: &Knapsack) -> u64;
 }
